The great format and clean up. IMPORTANT: Read below. All devs and contributors should use the provided formatter settings. If your IDE does not support the format, the rules are simple, it is the default java formatting rules, with 200char line lengths and 1 extra line after class definitions.

+++ src/main/java/com/cricketcraft/chisel/item/ItemChisel.java
@@ -193,7 +193,7 @@ public class ItemChisel extends Item implements IChiselMode {
 			y = buf.readInt();
 			z = buf.readInt();
 			block = buf.readInt();
-			meta = (byte) buf.readByte();
+			meta = buf.readByte();
 		}
 	}
